Behavioral/ + Core/ folder audit: fix 27 genuine issues across 43 findings
   CRITICAL (10):
   - BE-C1: KeEnterCriticalRegion for all 19 ERESOURCE acquire/release sites
   - BE-C2: CAS try-acquire in BeEngineGetChain/GetProcessContext (UAF race)
   - BE-C4: Reorder hash table cleanup before chain/context free in shutdown
   - PM-C1: Replace 42 ExAcquirePushLock with FltAcquirePushLock (critical region)
   - AD-C1: Remove 6 functions from PAGE section (use KSPIN_LOCK at DISPATCH)
   - FR-C1/C2: Wire PreSetInfo post-callback, fix draining IRQL handling

   HIGH (10):
   - DE-H1: Add ShuttingDown + rundown drain to CleanupByFlags failure path
   - DE-H2: Restore 15-field ETW schema for C2/Exfiltration/NetworkBlocked events
   - BE-H1: Add EX_RUNDOWN_REF to BehaviorEngine for shutdown safety
   - AD-H2/H3: Timeout check in shutdown waits, snapshot callbacks before invoke
   - TS-H1/H2: Division-by-zero guard, fix split-brain dual ThreatScoring
   - RE-H1/H2: Operator dispatch for BehaviorFlag/FileHash conditions
   - IOM-H2: Type-specific dispatch in IOCMatcher linear scan path

   MEDIUM (7):
   - DE-M1: Move power bridge unregister before BehaviorEngine shutdown
   - DE-M2: Fix g_InitFlags→InitFlags in CleanupByFlags (4 checks)
   - FR-M2: Add PostCreate.h to PostWrite.c/FileSystemCallbacks.c (struct mismatch)
   - FR-M3: Use pre-computed Length instead of strlen loop in IsScannable
   - IOM-M1: Hold shared lock during callback invocation (prevent UAF)
   - PM-M3: ReadNoFence for volatile RefCount in cleanup
   - ACT-M1: Register AttackChainTracker callback in BehaviorEngine init

+++ PhantomSensor/PhantomSensor/Callbacks/FileSystem/PostWrite.c
@@ -59,6 +59,7 @@
 #pragma warning(disable: 4324)  // structure was padded due to alignment specifier (fltKernel.h)
 #include "FileSystemCallbacks.h"
 #pragma warning(pop)
+#include "PostCreate.h"
 #include "../../Core/Globals.h"
 #include "../../Cache/ScanCache.h"
 #include "../../Shared/SharedDefs.h"
@@ -815,7 +816,7 @@ ShadowStrikePostWrite(
         // Detect full file overwrite pattern
         //
         if (writeContext.WriteOffset.QuadPart == 0 &&
-            writeContext.BytesWritten >= streamContext->ScanFileSize) {
+            (LONGLONG)writeContext.BytesWritten >= streamContext->ScanFileSize) {
             writeContext.IsFullOverwrite = TRUE;
         }
